# api-sql/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"getUsersResponse": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"users": {}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"getUserResponse": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"user": {}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"deleteUserResponse": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"deleteUserParams": {
				"properties": {
					"userId": {
						"type": "string"
					}
				},
				"required": [
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"updateUserResponse": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"user": {}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"updateUserParams": {
				"properties": {
					"avatar": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"cf_password": {
						"type": "string"
					}
				},
				"required": [
					"avatar",
					"name",
					"type",
					"password",
					"cf_password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"registerResponse": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"registerParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"account": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"loginResponse": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"access_token": {
						"type": "string"
					},
					"refresh_token": {
						"type": "string"
					},
					"user": {}
				},
				"required": [
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"loginParams": {
				"properties": {
					"account": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"account",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"activeResponse": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"activeParams": {
				"properties": {
					"active_token": {
						"type": "string"
					}
				},
				"required": [
					"active_token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"refreshTokenResponse": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"access_token": {
						"type": "string"
					},
					"user": {}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"googleLoginResponse": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"access_token": {
						"type": "string"
					},
					"refresh_token": {
						"type": "string"
					},
					"user": {},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"googleLoginParams": {
				"properties": {
					"id_token": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id_token",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"forgotPasswordResponse": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"forgotPasswordParams": {
				"properties": {
					"account": {
						"type": "string"
					}
				},
				"required": [
					"account"
				],
				"type": "object",
				"additionalProperties": false
			},
			"resetPasswordResponse": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"resetPasswordParams": {
				"properties": {
					"account": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"cf_password": {
						"type": "string"
					},
					"reset_token": {
						"type": "string"
					}
				},
				"required": [
					"account",
					"password",
					"cf_password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "node-server-ts-starter",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/admin/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Success ! here is the list of all users.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/getUsersResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/admin/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/getUserResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/deleteUserResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/deleteUserParams"
							}
						}
					}
				}
			}
		},
		"/api/admin/users/{userId}/edit": {
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/updateUserResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/updateUserParams"
							}
						}
					}
				}
			}
		},
		"/api/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/registerResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/registerParams"
							}
						}
					}
				}
			}
		},
		"/api/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/loginResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/loginParams"
							}
						}
					}
				}
			}
		},
		"/api/auth/active": {
			"post": {
				"operationId": "ActiveAccount",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/activeResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/activeParams"
							}
						}
					}
				}
			}
		},
		"/api/auth/logout": {
			"get": {
				"operationId": "Logout",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/auth/refresh_token": {
			"get": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/refreshTokenResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/auth/google_login": {
			"post": {
				"operationId": "GoogleLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/googleLoginResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/googleLoginParams"
							}
						}
					}
				}
			}
		},
		"/api/auth/forgot_password": {
			"post": {
				"operationId": "Forgot_password",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/forgotPasswordResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/forgotPasswordParams"
							}
						}
					}
				}
			}
		},
		"/api/auth/reset_password/{reset_token}": {
			"post": {
				"operationId": "Reset_password",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/resetPasswordResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/resetPasswordParams"
							}
						}
					}
				}
			}
		},
		"/api/user/me": {
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/updateUserResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/updateUserParams"
							}
						}
					}
				}
			}
		},
		"/api/user/reset_password": {
			"patch": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/resetPasswordResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/resetPasswordParams"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}